requirements.txt
# Step 2 - TTS + Speech Recognition (Improved)
gtts==2.4.0
pygame==2.5.2
pydantic==2.4.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pyaudio==0.2.14
webrtcvad==2.0.10

# Advanced Speech Recognition (Multilingual - Malayalam + English)
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3

# Step 3 - LLM Integration (Conversation Intelligence)
openai==1.12.0
anthropic==0.8.0

# Step 4 - Advanced TTS (Multilingual - Malayalam + English)
google-cloud-texttospeech==2.14.1
pydub==0.25.1

# Future additions:
# google-generativeai==0.3.0
# azure-cognitiveservices-speech==1.31.0
